fix wrong ActorSystem name when using MultiNodeClusterShardingSpec #28953
* use same testNameFromCallStack everywhere
