+++ akka-actor-testkit-typed/src/main/scala/akka/actor/testkit/typed/internal/TestKitUtils.scala
@@ -4,8 +4,6 @@
 
 package akka.actor.testkit.typed.internal
 
-import java.lang.reflect.Modifier
-
 import scala.util.control.Exception.Catcher
 
 import akka.actor.typed.scaladsl.Behaviors
@@ -72,48 +70,16 @@ private[akka] object TestKitUtils {
   // common internal utility impls for Java and Scala
   private val TestKitRegex = """akka\.testkit\.typed\.(?:javadsl|scaladsl)\.ActorTestKit(?:\$.*)?""".r
 
-  def testNameFromCallStack(classToStartFrom: Class[_]): String = {
-
-    def isAbstractClass(className: String): Boolean = {
-      try {
-        Modifier.isAbstract(Class.forName(className).getModifiers)
-      } catch {
-        case _: Throwable => false // yes catch everything, best effort check
-      }
-    }
-
-    val startFrom = classToStartFrom.getName
-    val filteredStack = Thread.currentThread.getStackTrace.iterator
-      .map(_.getClassName)
-      // drop until we find the first occurrence of classToStartFrom
-      .dropWhile(!_.startsWith(startFrom))
-      // then continue to the next entry after classToStartFrom that makes sense
-      .dropWhile {
-        case `startFrom`                            => true
-        case str if str.startsWith(startFrom + "$") => true // lambdas inside startFrom etc
-        case TestKitRegex()                         => true // testkit internals
-        case str if isAbstractClass(str)            => true
-        case _                                      => false
-      }
-
-    if (filteredStack.isEmpty)
-      throw new IllegalArgumentException(s"Couldn't find [${classToStartFrom.getName}] in call stack")
-
-    // sanitize for actor system name
-    scrubActorSystemName(filteredStack.next())
-  }
+  def testNameFromCallStack(classToStartFrom: Class[_]): String =
+    akka.testkit.TestKitUtils.testNameFromCallStack(classToStartFrom, TestKitRegex)
 
   /**
    * Sanitize the `name` to be used as valid actor system name by
    * replacing invalid characters. `name` may for example be a fully qualified
    * class name and then the short class name will be used.
    */
-  def scrubActorSystemName(name: String): String = {
-    name
-      .replaceFirst("""^.*\.""", "") // drop package name
-      .replaceAll("""\$\$?\w+""", "") // drop scala anonymous functions/classes
-      .replaceAll("[^a-zA-Z_0-9]", "_")
-  }
+  def scrubActorSystemName(name: String): String =
+    akka.testkit.TestKitUtils.scrubActorSystemName(name)
 
   def shutdown(system: ActorSystem[_], timeout: Duration, throwIfShutdownTimesOut: Boolean): Unit = {
     system.terminate()
