+++ akka-testkit/src/test/scala/akka/testkit/AkkaSpec.scala
@@ -4,8 +4,6 @@
 
 package akka.testkit
 
-import java.lang.reflect.Modifier
-
 import akka.actor.ActorSystem
 import akka.dispatch.Dispatchers
 import akka.event.Logging
@@ -43,54 +41,13 @@ object AkkaSpec {
           }
         }
       }
-                                                    """)
+      """)
 
   def mapToConfig(map: Map[String, Any]): Config = {
     import akka.util.ccompat.JavaConverters._
     ConfigFactory.parseMap(map.asJava)
   }
 
-  def testNameFromCallStack(classToStartFrom: Class[_]): String = {
-
-    def isAbstractClass(className: String): Boolean = {
-      try {
-        Modifier.isAbstract(Class.forName(className).getModifiers)
-      } catch {
-        case _: Throwable => false // yes catch everything, best effort check
-      }
-    }
-
-    val startFrom = classToStartFrom.getName
-    val filteredStack = Thread.currentThread.getStackTrace.iterator
-      .map(_.getClassName)
-      // drop until we find the first occurrence of classToStartFrom
-      .dropWhile(!_.startsWith(startFrom))
-      // then continue to the next entry after classToStartFrom that makes sense
-      .dropWhile {
-        case `startFrom`                            => true
-        case str if str.startsWith(startFrom + "$") => true // lambdas inside startFrom etc
-        case str if isAbstractClass(str)            => true
-        case _                                      => false
-      }
-
-    if (filteredStack.isEmpty)
-      throw new IllegalArgumentException(s"Couldn't find [${classToStartFrom.getName}] in call stack")
-
-    // sanitize for actor system name
-    scrubActorSystemName(filteredStack.next())
-  }
-
-  /**
-   * Sanitize the `name` to be used as valid actor system name by
-   * replacing invalid characters. `name` may for example be a fully qualified
-   * class name and then the short class name will be used.
-   */
-  def scrubActorSystemName(name: String): String = {
-    name
-      .replaceFirst("""^.*\.""", "") // drop package name
-      .replaceAll("""\$\$?\w+""", "") // drop scala anonymous functions/classes
-      .replaceAll("[^a-zA-Z_0-9]", "_")
-  }
 }
 
 abstract class AkkaSpec(_system: ActorSystem)
@@ -107,14 +64,14 @@ abstract class AkkaSpec(_system: ActorSystem)
   def this(config: Config) =
     this(
       ActorSystem(
-        AkkaSpec.testNameFromCallStack(classOf[AkkaSpec]),
+        TestKitUtils.testNameFromCallStack(classOf[AkkaSpec], "".r),
         ConfigFactory.load(config.withFallback(AkkaSpec.testConf))))
 
   def this(s: String) = this(ConfigFactory.parseString(s))
 
   def this(configMap: Map[String, _]) = this(AkkaSpec.mapToConfig(configMap))
 
-  def this() = this(ActorSystem(AkkaSpec.testNameFromCallStack(classOf[AkkaSpec]), AkkaSpec.testConf))
+  def this() = this(ActorSystem(TestKitUtils.testNameFromCallStack(classOf[AkkaSpec], "".r), AkkaSpec.testConf))
 
   val log: LoggingAdapter = Logging(system, this.getClass)
